# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/crops": {
            "get": {
                "tags": [
                    "Crops"
                ],
                "summary": "Get list of crops",
                "description": "Retrieve all available crops.",
                "operationId": "getCrops",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "crop_name": {
                                                "type": "string"
                                            },
                                            "crop_type": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Crops"
                ],
                "summary": "Create a new crop",
                "description": "Logs crop data in the database.",
                "operationId": "createCrop",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "farmer_id": {
                                        "type": "integer"
                                    },
                                    "crop_name": {
                                        "type": "string"
                                    },
                                    "crop_type": {
                                        "type": "string"
                                    },
                                    "average_amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "harvest_cycles": {
                                        "type": "integer"
                                    },
                                    "availability_start": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "availability_end": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "crop_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crop data logged successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/crops/store": {
            "post": {
                "tags": [
                    "Crops"
                ],
                "summary": "Store crop data",
                "description": "Stores crop data in the database.",
                "operationId": "storeCrop",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "farmer_id": {
                                        "type": "integer"
                                    },
                                    "crop_name": {
                                        "type": "string"
                                    },
                                    "crop_type": {
                                        "type": "string"
                                    },
                                    "average_amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "harvest_cycles": {
                                        "type": "integer"
                                    },
                                    "availability_start": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "availability_end": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "crop_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crop data stored successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/crops/{id}": {
            "get": {
                "tags": [
                    "Crops"
                ],
                "summary": "Get details of a specific crop",
                "description": "Retrieve detailed information about a specific crop.",
                "operationId": "getCrop",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the crop to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "crop_name": {
                                            "type": "string"
                                        },
                                        "crop_type": {
                                            "type": "string"
                                        },
                                        "price_per_kg": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Crops"
                ],
                "summary": "Update an existing crop",
                "description": "Updates crop details in the database.",
                "operationId": "updateCrop",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the crop to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "crop_name": {
                                        "type": "string"
                                    },
                                    "crop_type": {
                                        "type": "string"
                                    },
                                    "average_amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "harvest_cycles": {
                                        "type": "integer"
                                    },
                                    "price_per_kg": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "availability_start": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "availability_end": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "crop_image": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Crop updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Crop details updated successfully."
                                        },
                                        "crop_id": {
                                            "type": "integer",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation failed."
                                        },
                                        "details": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Crop not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Crop not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}